Add routes test
