+++ wcag-tool/backend/package.json
@@ -10,8 +10,8 @@
   "scripts": {
     "start": "node src/server.js",
     "dev": "nodemon src/server.js",
-    "test": "jest",
-    "lint": "eslint --ext .js,.jsx .",
+    "test:unit": "jest --testTimeout=10000 --detectOpenHandles",
+    "lint": "eslint --ext .js .",
     "lint:fix": "npm run lint -- --fix",
     "seed": "node src/utils/seed.js"
   },
@@ -32,9 +32,11 @@
     "eslint-plugin-jest": "^25.3.0",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^4.0.0",
-    "jest": "^27.3.1",
+    "jest": "^27.4.1",
+    "jest-extended": "^1.2.0",
     "mongodb-memory-server": "^8.0.2",
     "nodemon": "^2.0.15",
-    "prettier": "^2.4.1"
+    "prettier": "^2.4.1",
+    "supertest": "^6.1.6"
   }
 }
